Reworked sample users to resolve SonarCloud warnings
