+++ WordPress/src/main/java/org/wordpress/android/ui/selfhostedusers/UserListScreen.kt
@@ -219,7 +219,7 @@ private fun Screen(
 )
 fun UserListScreenPreview() {
     UserListScreen(
-        users = MutableStateFlow(sampleUserList).collectAsState(),
+        users = MutableStateFlow(SampleUsers.getSampleUsers()).collectAsState(),
         progressDialogState = null,
     )
 }
@@ -262,143 +262,3 @@ fun ProgressPreview() {
         progressDialogState = MutableStateFlow(progressDialogState).collectAsState(),
     )
 }
-
-private val sampleUser1 = UserWithEditContext(
-    id = 1,
-    username = "@sampleUser",
-    avatarUrls = emptyMap(),
-    capabilities = emptyMap(),
-    description = "User description",
-    email = "email@exmaple.com",
-    extraCapabilities = emptyMap(),
-    firstName = "Sample",
-    lastName = "User",
-    link = "example.com",
-    locale = "en-US",
-    name = "Sample User",
-    nickname = "User nickname",
-    registeredDate = "2023-01-01",
-    roles = listOf("admin"),
-    slug = "sample-user",
-    url = "example.com",
-)
-
-// TODO remove the avatar url
-private val sampleUser2 = UserWithEditContext(
-    id = 2,
-    username = "@sampleUserWithALongUserName",
-    avatarUrls = mapOf("sampleUserTwo" to "https://nickbradbury.com/wp-content/uploads/2022/03/1394-2.jpg"),
-    capabilities = emptyMap(),
-    description = "User description",
-    email = "email@exmaple.com",
-    extraCapabilities = emptyMap(),
-    firstName = "Sample",
-    lastName = "User",
-    link = "example.com",
-    locale = "en-US",
-    name = "Sample User",
-    nickname = "User nickname",
-    registeredDate = "2023-01-01",
-    roles = listOf("contributor"),
-    slug = "sample-user",
-    url = "example.com",
-)
-
-val sampleUserList = listOf(
-    sampleUser1,
-    sampleUser2,
-    sampleUser1,
-    sampleUser2,
-    sampleUser1,
-    sampleUser2,
-    sampleUser1,
-    sampleUser2,
-    sampleUser1,
-    sampleUser2,
-    sampleUser1,
-    sampleUser2,
-    sampleUser1,
-    sampleUser2,
-    sampleUser1,
-    sampleUser2,
-    sampleUser1,
-    sampleUser2,
-    sampleUser1,
-    sampleUser2,
-    sampleUser1,
-    sampleUser2,
-    sampleUser1,
-    sampleUser2,
-    sampleUser1,
-    sampleUser2,
-    sampleUser1,
-    sampleUser2,
-    sampleUser1,
-    sampleUser2,
-    sampleUser1,
-    sampleUser2,
-    sampleUser1,
-    sampleUser2,
-    sampleUser1,
-    sampleUser2,
-    sampleUser1,
-    sampleUser2,
-    sampleUser1,
-    sampleUser2,
-    sampleUser1,
-    sampleUser2,
-    sampleUser1,
-    sampleUser2,
-    sampleUser1,
-    sampleUser2,
-    sampleUser1,
-    sampleUser2,
-    sampleUser1,
-    sampleUser2,
-    sampleUser1,
-    sampleUser2,
-    sampleUser1,
-    sampleUser2,
-    sampleUser1,
-    sampleUser2,
-    sampleUser1,
-    sampleUser2,
-    sampleUser1,
-    sampleUser2,
-    sampleUser1,
-    sampleUser2,
-    sampleUser1,
-    sampleUser2,
-    sampleUser1,
-    sampleUser2,
-    sampleUser1,
-    sampleUser2,
-    sampleUser1,
-    sampleUser2,
-    sampleUser1,
-    sampleUser2,
-    sampleUser1,
-    sampleUser2,
-    sampleUser1,
-    sampleUser2,
-    sampleUser1,
-    sampleUser2,
-    sampleUser1,
-    sampleUser2,
-    sampleUser1,
-    sampleUser2,
-    sampleUser1,
-    sampleUser2,
-    sampleUser1,
-    sampleUser2,
-    sampleUser1,
-    sampleUser2,
-    sampleUser1,
-    sampleUser2,
-    sampleUser1,
-    sampleUser2,
-    sampleUser1,
-    sampleUser2,
-    sampleUser1,
-    sampleUser2,
-)
